refactor: 게임 시작 관련된 로직 변경
- ./Game/game.js로 이동!
